Exceptions

+++ spring-vscode/src/main/java/com/teste/spring_vscode/services/UserService.java
@@ -1,12 +1,17 @@
 package com.teste.spring_vscode.services;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.stereotype.Service;
+
 import com.teste.spring_vscode.entities.User;
 import com.teste.spring_vscode.repositories.UserRepository;
+import com.teste.spring_vscode.services.exceptions.DatabaseException;
+import com.teste.spring_vscode.services.exceptions.ResourceNotFoundException;
+
 
 @Service
 public class UserService {
@@ -20,7 +25,7 @@ public class UserService {
 
     public User findById(Long id) {
         Optional<User> obj = repository.findById(id);
-        return obj.get();
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public User insert(User obj) {
@@ -28,7 +33,16 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        Optional<User> user = repository.findById(id);
+		if (user.isPresent()) {
+			try {
+				repository.deleteById(id);
+			} catch (DataIntegrityViolationException e) {
+				throw new DatabaseException(e.getMessage());
+			}
+		} else {
+			throw new ResourceNotFoundException(id);
+		}
     }
 
     public User update(Long id, User obj) {
